WA-498 Apply standard --fix
