+++ app/index.js
@@ -9,19 +9,18 @@ const calculateInitialUrl = (account, safes) => {
   const validAccount = store.state.account.secondFA && Object.keys(account.secondFA).length > 0
   const validSafes = safes.safes && safes.safes.length > 0
 
-  if (validAccount && validSafes)
-    return ACCOUNT_URL
+  if (validAccount && validSafes) return ACCOUNT_URL
 
-  if (validAccount && !validSafes)
+  if (validAccount && !validSafes) {
     return {
       pathname: PASSWORD_URL,
       state: {
         dest: DOWNLOAD_APPS_URL
       }
     }
+  }
 
-  if (!validAccount && !validSafes)
-    return WELCOME_URL
+  if (!validAccount && !validSafes) return WELCOME_URL
 }
 
 store
@@ -32,7 +31,7 @@ store
       chrome.runtime.openOptionsPage
         ? chrome.runtime.openOptionsPage()
         : window.open(chrome.runtime.getURL('options/options.html'))
-  
+
       return
     }
